# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Proyectos",
        "contact": {
            "email": "andy.ancajima@mimco.com.pe"
        },
        "version": "1"
    },
    "paths": {
...
        "/GestionProyectos/public/index.php/vali_proy": {
            "post": {
                "tags": [
                    "Mantenimiento de Proyecto"
                ],
                "summary": "obtiene datos del proyecto",
                "operationId": "App\\Http\\Controllers\\GestionProyectosController::vali_proy",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "idProyecto que vamos ha seleccionar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "5"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "407": {
                        "description": "El IdProyecto ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/actu_proy": {
            "post": {
                "tags": [
                    "Mantenimiento de Proyecto"
                ],
                "summary": "Actualizar los datos del proyecto,solamente el atributo Alias",
                "operationId": "App\\Http\\Controllers\\GestionProyectosController::actu_proy",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "idProyecto que vamos ha seleccionar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "varAlias",
                        "in": "path",
                        "description": "Ingresar el Alias del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "usua_modi",
                        "in": "path",
                        "description": "El usuario que va realizar el actualizar tipo estructura",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "varAlias": {
                                        "type": "string"
                                    },
                                    "usua_modi": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "1000",
                                    "varAlias": "FERRETEROS",
                                    "usua_modi": "usuario"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizacion satisfactoria"
                    },
                    "407": {
                        "description": "El IdProyecto ingresado no se encuentra registrado."
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
        "/GestionProyectos/public/index.php/eleg_proc_estado": {
            "post": {
                "tags": [
                    "Mantenimiento de Proyecto"
                ],
                "summary": "obtenemos una lista de estado mediante el idProcEst",
                "operationId": "App\\Http\\Controllers\\GestionProyectosController::eleg_proc_estado",
                "parameters": [
                    {
                        "name": "intIdProcEsta",
                        "in": "path",
                        "description": "ingresamos el idProcEstado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProcEsta": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProcEsta": "4"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sin Mensaje"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
        "/GestionProyectos/public/index.php/gsp_proy_nom": {
            "post": {
                "tags": [
                    "Mantenimiento de Proyecto"
                ],
                "summary": "Obtener el nombre del programa + la zona",
                "operationId": "App\\Http\\Controllers\\GestionProyectosController::gsp_proy_nom",
                "parameters": [
                    {
                        "name": "intIdProy",
                        "in": "path",
                        "description": "ingresamos id del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "126"
                    },
                    {
                        "name": "intIdProyZona",
                        "in": "path",
                        "description": "ingresamos el proyecto zona",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "intIdProy": {
                                        "type": "string"
                                    },
                                    "intIdProyZona": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "intIdProy": "126",
                                    "intIdProyZona": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Obtener los datos mediante los parametros"
                    },
                    "404": {
                        "description": "No se encuentra el metodo"
                    }
                }
            }
        },
...
    }
}